requirements.txt
Flask==2.2.5
flask-wtf==1.0.1
reportlab==3.6.12
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==1.0.0
